packages/gateway/pyproject.toml
[project]
name = "loop-gateway"
version = "0.1.0"
description = "Loop LLM gateway: streaming providers, semantic cache, cost accounting, model aliases."
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
dependencies = [
    "loop",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["loop_gateway"]
